pyproject.toml
[tool.poetry]
name = "retail-intelligence-engine"
version = "0.1.0"
description = "Retail prediction backend with FastAPI"
authors = ["A Majumder"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1"
scikit-learn = "^1.3"
fastapi = "^0.110"
uvicorn = "^0.29"
pydantic = "^2.6"
tomli = "^2.0"
loguru = "^0.7"
joblib = "^1.3"

[tool.poetry.dev-dependencies]
pytest = "^8.1"
mypy = "^1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
